refactor: Extract DeleteOrgDialog and UpdateOrgForm
