+++ apps/web/src/features/organizations/components/OrgsSettings/UpdateOrgForm.tsx
@@ -1,0 +1,67 @@
+import { Button, TextField } from '@mui/material'
+import { FormProvider, useForm } from 'react-hook-form'
+import { showNotification } from '@/store/notificationsSlice'
+import {
+  type GetOrganizationResponse,
+  useOrganizationsUpdateV1Mutation,
+} from '@safe-global/store/gateway/AUTO_GENERATED/organizations'
+import { useAppDispatch } from '@/store'
+import { useCurrentOrgId } from '@/features/organizations/hooks/useCurrentOrgId'
+import { useIsAdmin } from '@/features/organizations/hooks/useOrgMembers'
+
+type UpdateOrganizationFormData = {
+  name: string
+}
+
+const UpdateOrgForm = ({ org }: { org: GetOrganizationResponse | undefined }) => {
+  const dispatch = useAppDispatch()
+  const [updateOrg] = useOrganizationsUpdateV1Mutation()
+  const orgId = useCurrentOrgId()
+  const isAdmin = useIsAdmin()
+
+  const formMethods = useForm<UpdateOrganizationFormData>({
+    mode: 'onChange',
+    values: {
+      name: org?.name || '',
+    },
+  })
+
+  const { register, handleSubmit, watch } = formMethods
+
+  const formName = watch('name')
+  const isNameChanged = formName !== org?.name
+
+  const onSubmit = handleSubmit(async (data) => {
+    try {
+      await updateOrg({ id: Number(orgId), updateOrganizationDto: { name: data.name } })
+
+      dispatch(
+        showNotification({
+          variant: 'success',
+          message: 'Successfully updated organization name',
+          groupKey: 'org-update-name',
+        }),
+      )
+    } catch (e) {
+      console.log(e)
+    }
+  })
+  return (
+    <FormProvider {...formMethods}>
+      <form onSubmit={onSubmit}>
+        <TextField
+          {...register('name')}
+          label="Organization name"
+          fullWidth
+          slotProps={{ inputLabel: { shrink: true } }}
+        />
+
+        <Button variant="contained" type="submit" sx={{ mt: 2 }} disabled={!isNameChanged || !isAdmin}>
+          Save
+        </Button>
+      </form>
+    </FormProvider>
+  )
+}
+
+export default UpdateOrgForm
